WeaponsApi/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "HoudiniTools-Api"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.105.0",
    "uvicorn==0.24.0",
    "pymongo==4.6.0",
    "motor==3.3.2",
    "python-multipart==0.0.6",
    "pydantic==2.11.4",
    "aiofiles==23.2.1",
    "python-dotenv==1.0.0",
] 